core/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-unfold==0.69.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.10
qrcode[pil]==8.0
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
